Refactor search functions
